Add support to shutdown KDE bundled to a single application package.
In this use case "kwinshutdown --watch <appname>" is installed as KDE
autostart service and watches registered services named on dbus. In case
the watched application has been exited, it shuts down the related KDE
session after the timout specified by --timeout. If not specified, the
timeout is set to 60 seconds by default.